Fixed compiler warning (there is no need for doubled in STK).
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@9839 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/modes/three_strikes_battle.hpp
@@ -56,7 +56,7 @@ private:
 
     float m_tire_radius[4];
     std::string m_tire_dir;
-    double m_tire_rotation;
+    float m_tire_rotation;
 
     PtrVector<PhysicalObject, REF> m_tires;
